+++ backend-rust/src/migrations.rs
@@ -201,16 +201,16 @@ pub enum SchemaVersion {
     RankingByLotteryPower,
     #[display("0012:Add removed bakers table")]
     TrackRemovedBakers,
-    #[display("0013:RankingByLotteryPower")]
-    BakerMetrics,
+    #[display("0013:Fix delegated_restake_earnings data in accounts")]
+    FixDelegatedStakeEarnings,
 }
 impl SchemaVersion {
     /// The minimum supported database schema version for the API.
     /// Fails at startup if any breaking database schema versions have been
     /// introduced since this version.
-    pub const API_SUPPORTED_SCHEMA_VERSION: SchemaVersion = SchemaVersion::BakerMetrics;
+    pub const API_SUPPORTED_SCHEMA_VERSION: SchemaVersion = SchemaVersion::TrackRemovedBakers;
     /// The latest known version of the schema.
-    const LATEST: SchemaVersion = SchemaVersion::BakerMetrics;
+    const LATEST: SchemaVersion = SchemaVersion::FixDelegatedStakeEarnings;
 
     /// Parse version number into a database schema version.
     /// None if the version is unknown.
@@ -238,7 +238,7 @@ impl SchemaVersion {
             SchemaVersion::DelegatedStakeCap => false,
             SchemaVersion::RankingByLotteryPower => false,
             SchemaVersion::TrackRemovedBakers => false,
-            SchemaVersion::BakerMetrics => false,
+            SchemaVersion::FixDelegatedStakeEarnings => false,
         }
     }
 
@@ -324,11 +324,13 @@ impl SchemaVersion {
             }
             SchemaVersion::TrackRemovedBakers => {
                 tx.as_mut()
-                    .execute(sqlx::raw_sql(include_str!("migrations/m0013-baker-metrics.sql")))
+                    .execute(sqlx::raw_sql(include_str!(
+                        "./migrations/m0013-fix-removed-delegators-restake.sql"
+                    )))
                     .await?;
-                SchemaVersion::BakerMetrics
+                SchemaVersion::FixDelegatedStakeEarnings
             }
-            SchemaVersion::BakerMetrics => unimplemented!(
+            SchemaVersion::FixDelegatedStakeEarnings => unimplemented!(
                 "No migration implemented for database schema version {}",
                 self.as_i64()
             ),